Project_1/Code_and_Experiments/Source_Code/requirements.txt
torch
torchvision
Pillow
fastapi
uvicorn
matplotlib
numpy
pip install opencv-python
